refactor: rename roles filename to Role
