+++ src/backend/riskManager/Risk.ts
@@ -1,6 +1,6 @@
 import { Expose, Transform, Type } from "class-transformer";
 import { DataType } from "../DataType";
-import { Role } from "../roles";
+import { Role } from "../Role";
 import { RiskLevel } from "./RiskLevel";
 
 /**
